requirements.txt
wget
tgcrypto
ffmpeg-python
py-tgcalls
pyrogram==2.0.59
yt-dlp
youtube-search-python
youtube-search
python-dotenv
dnspython
gitpython
asgiref
aiofiles
aiohttp
requests
pillow
heroku3
motor
psutil
future
wget
speedtest-cli
pymongo[srv]
pyromod~=1.5
googletrans==4.0.0-rc1
